AI_DD_Task_04_Streamlit/requirements.txt
PyPDF2
google-generativeai
streamlit
python-dotenv